Reduce to single history graph header
