+++ src/historygraph.h
@@ -9,11 +9,19 @@
 #include "global.h"
 #include "util.h"
 
+// for CJpegHistoryGraph
+#include <QImage>
+#include <QPainter>
+
+// for CSvgHistoryGraph
+#include <QXmlStreamWriter>
+#include <QXmlStreamAttributes>
+
 /* Definitions ****************************************************************/
 // number of history items to store
 #define NUM_ITEMS_HISTORY           600
 
-/* Interface ********************************************************************/
+/* Interface ******************************************************************/
 class AHistoryGraph
 {
 public:
@@ -95,6 +103,56 @@ protected:
     unsigned int iYSpace;
     QDate curDate;
     QTimer TimerDailyUpdate;
+};
+
+/* Implementations ************************************************************/
+class CJpegHistoryGraph : public QObject, virtual public AHistoryGraph
+{
+    Q_OBJECT
+
+public:
+    CJpegHistoryGraph();
+    virtual void Update ( );
+
+protected:
+    virtual void Save ( const QString sFileName );
+
+    virtual void rect ( const unsigned int x, const unsigned int y, const unsigned int width, const unsigned int height );
+    virtual void text ( const unsigned int x, const unsigned int y, const QString& value );
+    virtual void line ( const unsigned int x1, const unsigned int y1, const unsigned int x2, const unsigned int y2, const unsigned int strokeWidth = 1 );
+    virtual void point ( const unsigned int x, const unsigned int y, const unsigned int size, const QString& colour );
+
+private:
+    QImage PlotPixmap;
+    int iAxisFontWeight;
+
+public slots:
+    void OnTimerDailyUpdate() { Update(); }
+};
+
+class CSvgHistoryGraph : public QObject, virtual public AHistoryGraph
+{
+    Q_OBJECT
+
+public:
+    CSvgHistoryGraph();
+    virtual void Update();
+
+protected:
+    virtual void Save ( const QString sFileName );
+
+    virtual void rect ( const unsigned int x, const unsigned int y, const unsigned int width, const unsigned int height );
+    virtual void text ( const unsigned int x, const unsigned int y, const QString& value );
+    virtual void line ( const unsigned int x1, const unsigned int y1, const unsigned int x2, const unsigned int y2, const unsigned int strokeWidth = 1 );
+    virtual void point ( const unsigned int x, const unsigned int y, const unsigned int size, const QString& colour );
+
+private:
+    QXmlStreamAttributes svgRootAttributes;
+    QString svgImage;
+    QXmlStreamWriter svgStreamWriter;
+
+public slots:
+    void OnTimerDailyUpdate() { Update(); }
 };
 
 #endif // HISTORYGRAPH_H
